Adding Kinematic equations for 2 joint case.

+++ RoboticArm_Config.h
@@ -40,7 +40,7 @@
 namespace config
 {
     /* The physical length of each of the links in meters */
-    static constexpr float link_lengths[] = { 0.015, 0.015 };
+    static constexpr double link_lengths[] = { 0.015, 0.015 };
 
     /* Pair of pins used for these elements */
     static constexpr int quad_enc_pins[][2] = {{24, 25}, {26, 27}};
